Make deploy public api
