+++ lilac/cli.py
@@ -7,7 +7,7 @@ import click
 from . import __version__
 from .concepts.db_concept import DISK_CONCEPT_DB
 from .env import get_project_dir
-from .hf_docker_start import hf_docker_start as hf_docker_start_fn
+from .hf_docker_start import hf_docker_start
 from .load import load
 from .project import dir_is_project, init, project_dir_from_args
 from .server import start_server
@@ -89,9 +89,73 @@ def version() -> None:
 
 
 @click.command()
-def hf_docker_start() -> None:
+def hf_docker_start_command() -> None:
   """Prepares the binary by downloading datasets for the HuggingFace docker image."""
-  hf_docker_start_fn()
+  hf_docker_start()
+
+
+@click.command()
+@click.option(
+  '--project_dir',
+  help='The project directory to use for the demo. Defaults to `env.LILAC_PROJECT_DIR`.',
+  type=str,
+  required=True)
+@click.option(
+  '--hf_username', help='The huggingface username to use to authenticate for the space.', type=str)
+@click.option(
+  '--hf_space',
+  help='The huggingface space. Defaults to env.HF_STAGING_DEMO_REPO. '
+  'Should be formatted like `SPACE_ORG/SPACE_NAME`.',
+  type=str)
+@click.option('--dataset', help='The name of a dataset to upload', type=str, multiple=True)
+@click.option(
+  '--concept',
+  help='The name of a concept to upload. By default all lilac/ concepts are uploaded.',
+  type=str,
+  multiple=True)
+@click.option(
+  '--skip_build',
+  help='Skip building the web server TypeScript. '
+  'Useful to speed up the build if you are only changing python or data.',
+  type=bool,
+  is_flag=True,
+  default=False)
+@click.option(
+  '--skip_cache',
+  help='Skip uploading the cache files from .cache/lilac which contain cached concept pkl models.',
+  type=bool,
+  is_flag=True,
+  default=False)
+@click.option(
+  '--make_datasets_public',
+  help='When true, sets the huggingface datasets uploaded to public. Defaults to false.',
+  is_flag=True,
+  default=False)
+@click.option(
+  '--skip_data_upload',
+  help='When true, only uploads the wheel files without any other changes.',
+  is_flag=True,
+  default=False)
+@click.option(
+  '--use_pip',
+  help='When true, uses the public pip package. When false, builds and uses a local wheel.',
+  is_flag=True,
+  default=False)
+def deploy_hf_command(project_dir: str, hf_username: Optional[str], hf_space: Optional[str],
+                      dataset: list[str], concept: list[str], skip_build: bool, skip_cache: bool,
+                      make_datasets_public: bool, skip_data_upload: bool, use_pip: bool) -> None:
+  """Generate the huggingface space app."""
+  deploy_hf(
+    project_dir=project_dir,
+    hf_username=hf_username,
+    hf_space=hf_space,
+    datasets=dataset,
+    concepts=concept,
+    skip_build=skip_build,
+    skip_cache=skip_cache,
+    make_datasets_public=make_datasets_public,
+    skip_data_upload=skip_data_upload,
+    use_pip=use_pip)
 
 
 @click.command()
@@ -112,7 +176,8 @@ cli.add_command(init_command, name='init')
 cli.add_command(load_command, name='load')
 cli.add_command(start)
 
-cli.add_command(hf_docker_start)
+cli.add_command(deploy_hf_command, name='deploy_hf')
+cli.add_command(hf_docker_start_command, name='hf_docker_start')
 
 cli.add_command(concepts)
